make AuthorRow generic, replace other AuthorRows. Fix a few more height/position issues
