+++ packages/ui/src/components/ChatMessage/ChatMessageActions/MessageContainer.tsx
@@ -4,8 +4,8 @@ import { Dimensions } from 'react-native';
 
 import { ChatMessage } from '..';
 import { ScrollView, View } from '../../../core';
+import AuthorRow from '../../AuthorRow';
 import { NotebookPost } from '../../NotebookPost';
-import AuthorRow from '../AuthorRow';
 
 const MAX_MESSAGE_TO_SCREEN_RATIO = 0.3;
 const MAX_MESSAGE_TO_SCREEN_RATIO_NOTE = 0.5;
@@ -47,6 +47,7 @@ export function MessageContainer({
         author={post.author}
         authorId={post.authorId}
         sent={post.sentAt ?? 0}
+        type={post.type}
         // roles={roles}
       />
       <ChatMessage post={post} currentUserId={currentUserId} />
